chore(overview): rename and migrate useUpdateEditorEntities to useUpdateOverview
